requirements.txt
PyPDF2==3.0.1
pdf2docx==0.5.6
python-pptx==0.6.21
openpyxl==3.1.2
pdf2image==1.16.3
Pillow==10.0.0
pytesseract==0.3.10
reportlab==4.0.4
Flask==2.3.3
python-dotenv==1.0.0
Werkzeug==2.3.7 